Housekeeping Ability for Rasenbot #4 (#13)
* Removed auto assign and changed greetings.py to create a greetings message that requires new users to react to message to receive a role to access the server. Also changed formatting of some comments to make more readable

* Added new suggestions: Constant greetings message that requires new users to react to. Created constantvariables.py that contains variables that pertains to each server. Created a media folder full of copyright free images and gifs for use. Created the ability to unreact to welcome message and remove role

* Created more efficiency by only running the reaction add function if the message id is the one reacted to so it reduces the number of unnecessary checks and actions

* Added emoji to constantvariables, more efficiency for bot as it is able to remember welcome message after it goes offline

* Added a readme to show purpose of folder to developers

* Removed READ.txt

IMPORTANT:
Permissions must be set a certain way. (Might be best to add this to documentation for the bot)

DEFAULT PERMISSIONS for @everyone:
- everything must be turned off

DEFAULT PERMISSIONS for the role that is required to access server:
- able to view messages and send messages and connect to VC, etc.

PERMISSIONS FOR TEXT CHANNEL THAT SENDS GREETINGS MESSAGE for @everyone:
- able to view messages
- able to view message history
- everything else turned off

PERMISSIONS FOR ALL OTHER TEXT CHANNELS AND VC CHANNELS for @everyone:
- everything must be turned off

PERMISSIONS FOR ALL OTHER TEXT CHANNELS AND VC CHANNELS for role to access server:
- able to read and view message history
- able to connect to voices

Summary:
You receive role to access server if you react to greetings message and lose role if vice versa. Bot is able to remember greetings message so it will have access to it if bot goes offline.
This ability may get updated in the future if we want to change some of the rules or if we want to make the message more visually appealing or if we find more efficiency.